Create stacks (almost) empty, then update for the actual code.
The idea is to always be able to create new stacks, even if the template
and parameters are incorrect.

The stacks are created empty(they contain just a dummy resource), and
the actual code is deployed as an update that can be rolled back in case
of creation failure.

This allows much faster iterations while creating new stacks that may
fail to launch from the first attempt.

This fixes #1.